solve all (3)conflicts and refactore.
